Rename method to unzip
